Add files via upload
Here I have learnt about the use of grid placement property to make the grid items occupy the number of rows and columns of the grid and place the elements as per our requirement. I have also made use of flexbox inside the grid to align items at the centre.